requirements.txt
Flask
requests
pymongo
python-mimeparse
